Revert "Revert "Introduce state-specific employment accounts""
This reverts commit 4e7e59b3bac5b66ec0ce4d5795bf244a9f6bb219.

Reintroduce state-specific accounts for fiduciary obligations. After
some testing, I discovered that analytic accounts were not a good fit
for this because they are more intended for nominal accounts, not for
balance sheet liabilities, and designed for all analytic axes to be used
at once, not in a cherry-picking sort of way.